refactor: 상세화면 모듈화 작업
