+++ app/src/main/java/com/artem/animationjikan/domain/usecase/AnimationDetailUseCase.kt
@@ -3,7 +3,6 @@ package com.artem.animationjikan.domain.usecase
 import com.artem.animationjikan.data.mapper.toAnimationDetailEntity
 import com.artem.animationjikan.domain.entities.DetailEntity
 import com.artem.animationjikan.domain.repository.AnimationRepository
-import com.artem.animationjikan.domain.repository.MangaRepository
 import javax.inject.Inject
 
 class AnimationDetailUseCase @Inject constructor(
